fix(codacy): add CLSCompliant attributes to all assemblies
- Add [assembly: CLSCompliant(true)] to all 11 projects
- Fix MegEntry.Crc32 (uint → long) and Flags (ushort → int) for CLS compliance
- Upgrade coverlet to 8.0.0 (6.0.5 doesn't exist, jumped from 6.0.4 to 8.0.0)

Addresses 233 Codacy "Mark Assemblies as CLS Compliant" findings.

+++ src/SwfocTrainer.Meg/MegEntry.cs
@@ -2,8 +2,8 @@ namespace SwfocTrainer.Meg;
 
 public sealed record MegEntry(
     string Path,
-    uint Crc32,
+    long Crc32,
     int Index,
     int SizeBytes,
     int StartOffset,
-    ushort Flags = 0);
+    int Flags = 0);
